# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/google/login": {
      "get": {
        "operationId": "OAuthController_initiateGoogleSignIn",
        "parameters": [
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorization URL generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "authorizationUrl": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed to generate authorization URL"
          }
        },
        "summary": "Initiate Google OAuth sign-in",
        "tags": [
          "oauth"
        ]
      }
    },
    "/api/auth/google/callback": {
      "get": {
        "operationId": "OAuthController_completeGoogleSignIn",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OAuth sign-in completed successfully"
          },
          "400": {
            "description": "Invalid OAuth callback"
          }
        },
        "summary": "Complete Google OAuth sign-in",
        "tags": [
          "oauth"
        ]
      }
    },
    "/api/auth/discord/login": {
      "get": {
        "operationId": "OAuthController_initiateDiscordSignIn",
        "parameters": [
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorization URL generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "authorizationUrl": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed to generate authorization URL"
          }
        },
        "summary": "Initiate Discord OAuth sign-in",
        "tags": [
          "oauth"
        ]
      }
    },
    "/api/auth/discord/callback": {
      "get": {
        "operationId": "OAuthController_completeDiscordSignIn",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OAuth sign-in completed successfully"
          },
          "400": {
            "description": "Invalid OAuth callback"
          }
        },
        "summary": "Complete Discord OAuth sign-in",
        "tags": [
          "oauth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lazy Map API",
    "description": "API for generating battle maps",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}